AI Assistant/requirements.txt
aiohttp==3.9.1
customtkinter==5.2.2
openai>=1.6.1
httpx>=0.26.0
Pillow==10.1.0
tqdm==4.66.1
darkdetect==0.8.0
packaging==24.1
